backend/requirements.txt
flask
flask-cors
openai
python-dotenv
python-pptx
Pillow
requests